Update - split structure to entities modules
